aider: Increase size of buffer and make plugin configurable to open with vsplit, hsplit, or floating buffer.
# Context:
USER: Please go back to using the vsplit method for the plugin
ASSISTANT: To revert back to opening the terminal in a vertical split, you'll need to modify the `OpenAider`
^C KeyboardInterrupt
USER: please increase the size of the buffer so there is less space around it and also make the plugin configurable soit can be opened with a vsplit, hsplit or floating buffer
ASSISTANT: To make the plugin configurable and to increase the size of the buffer, you'll need to modify the `OpenAider` function in your `plugin/aider.vim` file.

Here's how you can modify the function:

```vim
plugin/aider.vim
<<<<<<< ORIGINAL
    function! OpenAider()
        " Create a new buffer for the terminal
        let l:buf = nvim_create_buf(v:false, v:true)
        " Calculate the size and position of the floating window
        let l:width = 80
        let l:height = 24
        let l:row = (nvim_win_get_height(0) - l:height) / 2
        let l:col = (nvim_win_get_width(0) - l:width) / 2
        " Create a new floating window for the terminal
        let l:win = nvim_open_win(l:buf, v:true, {'relative': 'editor', 'width': l:width, 'height': l:height, 'row': l:row, 'col': l:col})
        " Run 'aider' in the terminal
        call termopen('aider', {'on_exit': function('s:OnExit')})
        " Make the terminal window active
        call nvim_set_current_win(l:win)
    endfunction
=======
    function! OpenAider()
        " Create a new buffer for the terminal
        let l:buf = nvim_create_buf(v:false, v:true)
        " Get the user's preferred window type
        let l:window_type = get(g:, 'aider_window_type', 'vsplit')
        " Open the terminal in the preferred window type
        if l:window_type == 'vsplit'
            vnew | terminal
        elseif l:window_type == 'hsplit'
            new | terminal
        else
            " Calculate the size and position of the floating window
            let l:width = nvim_win_get_width(0) - 4
            let l:height = nvim_win_get_height(0) - 4
            let l:row = 2
            let l:col = 2
            " Create a new floating window for the terminal
            let l:win = nvim_open_win(l:buf, v:true, {'relative': 'editor', 'width': l:width, 'height': l:height, 'row': l:row, 'col': l:col})
        endif
        " Run 'aider' in the terminal
        call termopen('aider', {'on_exit': function('s:OnExit')})
        " Make the terminal window active
        call nvim_set_current_win(l:win)
    endfunction
>>>>>>> UPDATED
```

This code gets the user's preferred window type from the `g:aider_window_type` variable. If the variable is not set, it defaults to 'vsplit'. Then, it opens the terminal in the preferred window type. If the preferred window type is a floating window, it calculates the size and position of the window to be slightly smaller than the Neovim window.

Users can set their preferred window type in their `init.vim` or `init.lua` file like this:

```vim
let g:aider_window_type = 'vsplit'  " or 'hsplit' or 'float'
```

Please note that you need to replace `aider` with the actual command you want to run in the terminal. If `aider` is a script, make sure it's in your PATH. If it's a function defined in your `init.vim` file, you'll need to call it with `:call aider()` instead of `aider`.