aider: Replaced 'aider' command with 'echo' command in the `s:RunAider` function.
# Context:
USER: I did `git reset --hard HEAD~1` to discard the last edits.
ASSISTANT: Okay, that command will reset your Git repository to the state it was in before the most recent commit. This means the changes you made in the last commit have been discarded. If you need to make any further changes or need assistance with anything else, feel free to ask.
USER: that didn't fix it - I am seeing this in the terminal
ASSISTANT: I'm sorry, but I can't see the error message or the output you're referring to. Could you please provide more details or the error message you're seeing in the terminal?
USER: zsh: bad option: -+

[Process exited 1]
ASSISTANT: The error message "zsh: bad option: -+" suggests that the shell (zsh) is receiving an invalid option. This is likely due to a command being passed to the shell that includes "-+" which is not a valid option for most commands.

The issue seems to be with the `termopen` function call in the `s:RunAider` function. The `termopen` function is trying to execute the 'aider' command, but it seems like the 'aider' command is not recognized or not working as expected.

To debug this issue, you can replace the 'aider' command with a command that you know works (like 'echo') to see if the problem is with the 'aider' command or with the `termopen` function itself.

Here's how you can do it:

```vim
plugin/aider.vim
<<<<<<< ORIGINAL
    function! s:RunAider()
        " Run 'aider' in the terminal
        call termopen('aider', {'on_exit': function('s:OnExit')})
    endfunction
=======
    function! s:RunAider()
        " Run 'echo' in the terminal
        call termopen('echo "Hello, World!"', {'on_exit': function('s:OnExit')})
    endfunction
>>>>>>> UPDATED
```

This change will make the `s:RunAider` function run the 'echo' command instead of the 'aider' command. If this works, then the problem is likely with the 'aider' command. If it still doesn't work, then the problem is likely with the `termopen` function or the way it's being used.