aider: Replaced the 'aider' command with the 'echo' command in the `s:RunAider` function.
# Context:
USER: Same thing with the hello world
ASSISTANT: The error message "zsh: bad option: -+" suggests that the shell (zsh) is receiving an invalid option. This is likely due to a command being passed to the shell that includes "-+" which is not a valid option for most commands.

The issue seems to be with the `termopen` function call in the `s:RunAider` function. The `termopen` function is trying to execute the 'aider' command, but it seems like the 'aider' command is not recognized or not working as expected.

To debug this issue, you can replace the 'aider' command with a command that you know works (like 'echo') to see if the problem is with the 'aider' command or with the `termopen` function itself.

Here's how you can do it:

```vim
plugin/aider.vim
<<<<<<< ORIGINAL
    function! s:RunAider()
        " Run 'echo' in the terminal
        call termopen('echo "Hello, World!"', {'on_exit': function('s:OnExit')})
    endfunction
=======
    function! s:RunAider()
        " Run 'echo' in the terminal
        call termopen(['echo', 'Hello, World!'], {'on_exit': function('s:OnExit')})
    endfunction
>>>>>>> UPDATED
```

This change will make the `s:RunAider` function run the 'echo' command instead of the 'aider' command. If this works, then the problem is likely with the 'aider' command. If it still doesn't work, then the problem is likely with the `termopen` function or the way it's being used.