aider: Fix issue with opening terminal in a modified buffer by checking if the buffer is saved before opening the terminal, and switch to the new window before opening the terminal.
# Context:
USER: Error detected while processing function OpenAider[6]..<SNR>30_RunAider:
line    2:
Can only call this function in an unmodified buffer
Press ENTER or type command to continue
---
ASSISTANT: The error message suggests that the `termopen` function is being called in a buffer that has unsaved changes. The `termopen` function can only be called in an unmodified buffer in Neovim.

To fix this, you can add a check at the beginning of the `OpenAider` function to ensure that the current buffer is saved before opening the terminal. If the buffer is not saved, you can display a message to the user to save the buffer first.

Here's how you can do it:

```vim
plugin/aider.vim
<<<<<<< ORIGINAL
    function! OpenAider()
        " Get the user's preferred window type
        let l:window_type = get(g:, 'aider_window_type',
^C KeyboardInterrupt
USER: it used to work when we had a single float window
ASSISTANT: I see. The issue might be due to the fact that we are trying to open a terminal in a new window before switching to that window. The `termopen` function is likely trying to open the terminal in the current window, which is not the new window we just created.

To fix this, we can modify the `s:OpenWindow` function to switch to the new window before opening the terminal. Here's how you can do it:

```vim
plugin/aider.vim
<<<<<<< ORIGINAL
    function! s:OpenWindow(window_type)
        " Create a new buffer for the terminal
        let l:buf = nvim_create_buf(v:false, v:true)
        " Open the terminal in the preferred window type
        if a:window_type == 'vsplit'
            vnew | terminal
        elseif a:window_type == 'hsplit'
            new | terminal
        else
            " Calculate the size and position of the floating window
            let l:width = nvim_win_get_width(0) - 4
            let l:height = nvim_win_get_height(0) - 4
            let l:row = 2
            let l:col = 2
            " Create a new floating window for the terminal
            let l:win = nvim_open_win(l:buf, v:true, {'relative': 'editor', 'width': l:width, 'height': l:height, 'row': l:row, 'col': l:col})
            " Make the terminal window active
            call nvim_set_current_win(l:win)
        endif
    endfunction
=======
    function! s:OpenWindow(window_type)
        " Create a new buffer for the terminal
        let l:buf = nvim_create_buf(v:false, v:true)
        " Open the terminal in the preferred window type
        if a:window_type == 'vsplit'
            vnew | call setbufvar(l:buf, '&buftype', 'nofile') | execute 'terminal'
        elseif a:window_type == 'hsplit'
            new | call setbufvar(l:buf, '&buftype', 'nofile') | execute 'terminal'
        else
            " Calculate the size and position of the floating window
            let l:width = nvim_win_get_width(0) - 4
            let l:height = nvim_win_get_height(0) - 4
            let l:row = 2
            let l:col = 2
            " Create a new floating window for the terminal
            let l:win = nvim_open_win(l:buf, v:true, {'relative': 'editor', 'width': l:width, 'height': l:height, 'row': l:row, 'col': l:col})
            " Make the terminal window active
            call nvim_set_current_win(l:win)
            " Set the buffer type to 'nofile' and open the terminal
            call setbufvar(l:buf, '&buftype', 'nofile') | execute 'terminal'
        endif
    endfunction
>>>>>>> UPDATED
```

This change will make the `s:OpenWindow` function switch to the new window before opening the terminal. This should fix the issue of the `termopen` function trying to open the terminal in the current window.