aider: Update `OpenAider` function to accept an optional `window_type` argument.
# Context:
USER: please update the openaider function so it can take an optional window_type argument with the same options as the g variable
ASSISTANT: To make the necessary changes, I'll need write access to the `plugin/aider.vim` file. Could you please make it *read-write*?
USER: I added these *read-write* files: plugin/aider.vim
ASSISTANT: Great! Here's how we can modify the `OpenAider` function to accept an optional `window_type` argument. If the argument is not provided, it will default to the value of the `g:aider_window_type` variable.

```vim
plugin/aider.vim
<<<<<<< ORIGINAL
    function! OpenAider()
        " Get the user's preferred window type
        let l:window_type = get(g:, 'aider_window_type', 'vsplit')
        " Open the window
        call s:OpenWindow(l:window_type)
        " Run 'aider'
        call s:RunAider()
    endfunction
=======
    function! OpenAider(window_type)
        " Get the user's preferred window type
        let l:window_type = get(a:, 'window_type', get(g:, 'aider_window_type', 'vsplit'))
        " Open the window
        call s:OpenWindow(l:window_type)
        " Run 'aider'
        call s:RunAider()
    endfunction
>>>>>>> UPDATED
```

This change makes `OpenAider` accept an optional `window_type` argument. If the argument is not provided, it defaults to the value of the `g:aider_window_type` variable, or 'vsplit' if `g:aider_window_type` is not set.