aider: Make the window_type variable optional.
# Context:
USER: Please make the window_type variable optional
ASSISTANT: To make the `window_type` variable optional in the `OpenAider` function, we can modify the function definition to include a default value for `window_type`. Here's how to do it:

```vim
plugin/aider.vim
<<<<<<< ORIGINAL
    function! OpenAider(window_type)
        " Get the user's preferred window type
        let l:window_type = get(a:, 'window_type', get(g:, 'aider_window_type', 'vsplit'))
        " Open the window
        call s:OpenWindow(l:window_type)
        " Run 'aider'
        call s:RunAider()
    endfunction
=======
    function! OpenAider(window_type = '')
        " Get the user's preferred window type
        let l:window_type = get(a:, 'window_type', get(g:, 'aider_window_type', 'vsplit'))
        " Open the window
        call s:OpenWindow(l:window_type)
        " Run 'aider'
        call s:RunAider()
    endfunction
>>>>>>> UPDATED
```

This change makes `window_type` an optional argument in the `OpenAider` function. If the argument is not provided, it defaults to an empty string, which then causes the `get` function to use the value of `g:aider_window_type` or 'vsplit' if `g:aider_window_type` is not set.