aider: Added instructions to the readme about the arguments that can be passed to the OpenAider function.
# Context:
USER: Great, nearly there. Please add instructions to the readme about the arguments that can be passed to the OpenAider function
ASSISTANT: To add instructions about the arguments that can be passed to the OpenAider function, we need to modify the README.md file. Here are the changes:

```markdown
README.md
<<<<<<< ORIGINAL
## Usage

Press `<leader><Space><Space>` to open a terminal window with the Aider command.

Press `<leader><Space>3` to open a terminal window with the Aider command using the gpt-3.5-turbo-16k model.
=======
## Usage

The Aider Plugin for Neovim provides the `OpenAider` function, which you can call to open a terminal window with the Aider command. The `OpenAider` function accepts the following arguments:

- `model`: The name of the model to use. Default is `gpt-3.5-turbo-16k`.
- `prompt`: The initial prompt to pass to the Aider command. Default is an empty string.

Here are some examples of how to use the `OpenAider` function:

```vim
:OpenAider gpt-3.5-turbo-16k "Hello, Aider!"
:OpenAider gpt-3.5-turbo-16k
:OpenAider
```

You can also use the following keybindings:

- Press `<leader><Space><Space>` to open a terminal window with the Aider command.
- Press `<leader><Space>3` to open a terminal window with the Aider command using the gpt-3.5-turbo-16k model.
>>>>>>> UPDATED
```

Please let me know if you need any further changes.