aider: Updated the `CHANGELOG.md` file with the correct date.
# Context:
USER: It is 2023-07-12
ASSISTANT: Great, let's update the `CHANGELOG.md` file with the correct date.

```markdown
newfile:CHANGELOG.md
<<<<<<< ORIGINAL
# Changelog

All notable changes to this project will be documented in this file.

## [0.1] - YYYY-MM-DD

### Added
- Initial release
=======
# Changelog

All notable changes to this project will be documented in this file.

## [0.1] - 2023-07-12

### Added
- Initial release
>>>>>>> UPDATED
```

Please replace the content of `CHANGELOG.md` with the updated version. After that, you can commit the changes.