aider: Modified `plugin/aider.lua` to export the `OpenAider` function by creating a new table `M` and returning it at the end of the module.
# Context:
USER: When I run the sample lua config it says module aider not found - is it being exported correctly from the plugin?
ASSISTANT: To verify if the module is being exported correctly, we need to check the `plugin/aider.lua` file. However, this file is currently read-only. Could you please make it read-write so that we can inspect and possibly modify it?
USER: I added these *read-write* files: plugin/aider.lua
ASSISTANT: Great, now that we have write access to `plugin/aider.lua`, let's check if the `OpenAider` function is being exported correctly.

In Lua, to make a function available to other modules, we need to return it at the end of the module. Currently, the `OpenAider` function is not being returned, which is why you're getting a "module 'aider' not found" error.

Here's how to modify `plugin/aider.lua` to export the `OpenAider` function:

```lua
plugin/aider.lua
<<<<<<< ORIGINAL
function OpenAider(command, window_type)
    command = command or 'aider'
    window_type = window_type or 'vsplit'
    open_window(window_type)
    vim.fn.termopen(command, {on_exit = 'OnExit'})
end
=======
local M = {}

function M.OpenAider(command, window_type)
    command = command or 'aider'
    window_type = window_type or 'vsplit'
    open_window(window_type)
    vim.fn.termopen(command, {on_exit = 'OnExit'})
end

return M
>>>>>>> UPDATED
```

This change creates a new table `M` and adds the `OpenAider` function to it. Then it returns `M` at the end of the module, which makes the `OpenAider` function available to other modules. Now you should be able to require the 'aider' module and call `aider.OpenAider()`.