aider: Added content to the `aider.txt` help file for the `aide.nvim` plugin.
# Context:
USER: I would like to make a help file so that :help aider.nvim shows the information about the function and keybindings
ASSISTANT: To create a help file for your Neovim plugin, you will need to create a new file in the `doc` directory of your plugin. This file should be named after your plugin with a `.txt` extension. In your case, the file would be `aider.txt`.

The help file should contain a modeline at the bottom to set the correct options for viewing the help file. It should also contain tags for each section so that users can jump directly to a section using the `:help` command.

Here's a basic structure for your help file:

```vim
* aide.txt* For aide.nvim plugin

1. Introduction         |aide-introduction|
2. Functions            |aide-functions|
3. Keybindings          |aide-keybindings|

==============================================================================
1. Introduction                                             *aide-introduction*

This is the introduction.

==============================================================================
2. Functions                                                *aide-functions*

This is the functions.

==============================================================================
3. Keybindings                                              *aide-keybindings*

This is the keybindings.

vim:tw=78:ts=8:ft=help:norl:
```

You can replace "This is the introduction.", "This is the functions.", and "This is the keybindings." with the actual content.

Here's the *edit block* for the new file:

```vim
doc/aider.txt
<<<<<<< ORIGINAL

=======
* aide.txt* For aide.nvim plugin

1. Introduction         |aide-introduction|
2. Functions            |aide-functions|
3. Keybindings          |aide-keybindings|

==============================================================================
1. Introduction                                             *aide-introduction*

This is the introduction.

==============================================================================
2. Functions                                                *aide-functions*

This is the functions.

==============================================================================
3. Keybindings                                              *aide-keybindings*

This is the keybindings.

vim:tw=78:ts=8:ft=help:norl:
>>>>>>> UPDATED
```

Please replace the placeholders with the actual content.