aider: Added an autocomplete tag for `aider.nvim` in the `aider.txt` file.
# Context:
USER: do i need to do anything to rebuild the help autocompletes? I'm still seeing the old ones even though the doc file has been updated
ASSISTANT: Yes, you need to rebuild the help tags in Vim or Neovim after updating the help files. This is done with the `:helptags` command.

You need to run this command in the directory where your help files are located. If your help files are in the `doc` directory, you would run `:helptags doc/`.

Please note that you need to have write permissions to the directory where your help files are located to run the `:helptags` command. If you don't have write permissions, you will need to change the permissions or run Vim or Neovim with sudo.

Here is the command you need to run in Vim or Neovim:

```vim
:helptags doc/
```

After running this command, the help autocompletes should be updated.
USER: can i specificy a plugin instead of a folder?
ASSISTANT: No, the `:helptags` command in Vim and Neovim does not accept a plugin as an argument. It only accepts a directory path.

The directory should contain the documentation files for the plugin. In most cases, this is the `doc` directory inside the plugin's directory.

If you're using a plugin manager like vim-plug, Vundle, or Pathogen, they usually provide a command to generate help tags for all installed plugins. For example, with vim-plug, you can use `:PlugInstall` or `:PlugUpdate` to update the plugins and generate the help tags.

If you're not using a plugin manager, you'll need to manually run `:helptags` on the `doc` directory of each plugin.
USER: what about packer
ASSISTANT: If you're using `packer.nvim` as your plugin manager, it should automatically generate the help tags for all plugins when they are installed or updated.

However, if you need to manually regenerate the help tags, you can use the `PackerCompile` command. This command compiles all your plugins and regenerates the help tags.

Here is the command you need to run in Neovim:

```vim
:PackerCompile
```

After running this command, the help autocompletes should be updated.
USER: Now I have no autocomplete - please add just one autocomplete for aider.nvim
ASSISTANT: To add an autocomplete for `aider.nvim`, we need to add a tag in the `aider.txt` file. This tag will be used for autocompletion when you access the help.

Here are the changes:

```vim
doc/aider.txt
<<<<<<< ORIGINAL
* aider.txt* For aider.nvim plugin

1. Introduction
2. Functions
3. Keybindings

==============================================================================
1. Introduction

Aider is a simple plugin for Neovim that allows you to open a terminal window inside Neovim and run the Aider.

==============================================================================
2. Functions

The Aider Plugin for Neovim provides the `OpenAider` function, which you can call to open a terminal window with the Aider command. The `OpenAider` function accepts the following arguments:

- `command`: The full aider command to use - defaults to `a
^C KeyboardInterrupt
USER: Please don't regenerate the whole doc file, just give me hte parts that need to chnage
ASSISTANT: Sure, here's the change needed to add an autocomplete for `aider.nvim`:

```vim
doc/aider.txt
<<<<<<< ORIGINAL
* aider.txt* For aider.nvim plugin
=======
* aider.txt* For aider.nvim plugin *aider-nvim*
>>>>>>> UPDATED
```

This change adds a tag `*aider-nvim*` to the first line of the `aider.txt` file. Now, when you type `:help aider.nvim` in Neovim, it will autocomplete to this tag.