aider: Added release entry for version 2.0 with new features and improvements.
USER: please add a new entry int he changelog for releasing v2 and put in a summary of the changes

ASSISTANT: Sure, I can add a new entry to the CHANGELOG.md file for the release of version 2.0.

Here's the proposed change:

```markdown
CHANGELOG.md
<<<<<<< ORIGINAL

All notable changes to this project will be documented in this file.

- Initial release
=======

All notable changes to this project will be documented in this file.

- Aider now automatically adds all open buffers when it opens.
- Added tips for working with buffers in Vim to the README.md file.

- Initial release
>>>>>>> UPDATED
```

Please confirm if this change is acceptable.

fixing version number